chatbot-backend/pyproject.toml
[project]
name = "chatbot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.29.0",
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.109.0",
    "httpx>=0.26.0",
    "markdown>=3.5.2",
    "openai>=1.12.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.7.1",
    "sentence-transformers>=2.3.1",
    "sqlalchemy[asyncio]>=2.0.25",
    "tiktoken>=0.5.2",
    "uvicorn[standard]>=0.27.0",
]
